Add routing
* add api gateway
* add product service

+++ db-service/src/main/java/com/istore/db/service/repository/OrderRepository.java
@@ -1,10 +1,9 @@
 package com.istore.db.service.repository;
 
-
 import com.istore.db.service.model.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderRepository extends JpaRepository<Order, Integer> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
 }
