+++ db-service/src/main/java/com/istore/db/service/repository/ProductRepository.java
@@ -1,10 +1,9 @@
 package com.istore.db.service.repository;
 
-
 import com.istore.db.service.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 }
